+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - [#12](https://github.com/oleasteo/koreader-screenlockpin/issues/12)
   A notification is shown after unlock, if PIN attempts failed too many times
   (throttled at least twice) since locked.
+- Menu toggle to disable the lock screen.
 
 ## [2025.11] - 2025-11-04